+++ packages/eyes-sdk-core/lib/errors/NewTestError.js
@@ -1,24 +1,12 @@
-'use strict'
-const TestFailedError = require('./TestFailedError')
-const SessionStartInfo = require('../server/SessionStartInfo')
+const EyesError = require('./EyesError')
 
 /**
  * Indicates that a new test (i.e., a test for which no baseline exists) ended.
  */
-class NewTestError extends TestFailedError {
-  /**
-   * Creates a new NewTestError instance.
-   *
-   * @param {TestResults} testResults - The results of the current test if available, {@code null} otherwise.
-   * @param {string|SessionStartInfo} messageOrSession - The error description or SessionStartInfo with test details.
-   */
-  constructor(testResults, messageOrSession) {
-    if (messageOrSession instanceof SessionStartInfo) {
-      const testName = `'${messageOrSession.getScenarioIdOrName()}' of '${messageOrSession.getAppIdOrName()}'`
-      messageOrSession = `${testName}. Please approve the new baseline at ${testResults.getUrl()}`
-    }
-
-    super(testResults, messageOrSession)
+class NewTestError extends EyesError {
+  constructor(testResult) {
+    const message = `Test '${testResult.name}' of '${testResult.appName}' is new! Please approve the new baseline at ${testResult.url}`
+    super(message, {reason: 'test new', testResult})
   }
 }