+++ packages/types/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- add support for `throwErr` argument to `close` and `closeAllEyes`
+- return array of tests results from `close` and `abort`
 
 ## 1.0.6 - 2021/8/3